requirements.txt
torch
torchvision
pywebview
Pillow
insightface
networkx
numpy
opencv-python
pystray
lmdb
pillow-heif
